chore: rename project and add crewai test support
