+++ src/auto_k8s_pilot/config/layers/crew-infra-cluster.yaml
