+++ src/auto_k8s_pilot/config/tasks.yaml
